task6 inprogree

+++ task6/include/curve.h
@@ -5,7 +5,7 @@
 
 float smooth_curve(float * sample, int n,int smooth_factor, float * smooth);
 
-float amplitude(int * input, int m,int Amp);
+float amplitude(float * input, int m,float Amp);
 
 int get_data(FILE* datafile, float* data, int sz_data);
